improvements - restricting category removal

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -116,9 +116,19 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
+        $products = $this->repository->productsByCategoryId($id);
+
+        if (count($products) > 0) {
+            return redirect()
+                ->route('categories.index')
+                ->with('message', 'Não pode deletar porque existem produtos vinculados a essa categoria!');
+        }
+
         $this->repository->delete($id);
 
-        return redirect()->route('categories.index');
+        return redirect()
+            ->route('categories.index')
+            ->with('success', 'Categoria deletada com sucesso!');
     }
 
     public function search(Request $request)
